Make apply filter button works on embeddings
