+++ frontend/src/metabase/dashboard/containers/DashboardApp.jsx
@@ -33,7 +33,6 @@ import * as Urls from "metabase/lib/urls";
 
 import Dashboards from "metabase/entities/dashboards";
 
-import { isNotNull } from "metabase/core/utils/types";
 import * as dashboardActions from "../actions";
 import {
   getIsEditing,
@@ -59,7 +58,6 @@ import {
   getIsHeaderVisible,
   getIsAdditionalInfoVisible,
 } from "../selectors";
-import { ApplyButton } from "./DashboardApp.styled";
 
 function getDashboardId({ dashboardId, params }) {
   if (dashboardId) {
@@ -161,31 +159,11 @@ const DashboardApp = props => {
     setIsShowingToaster(false);
   }, []);
 
-  const { parameterValues, draftParameterValues, setParameterValues } = props;
-  const hasUnappliedFilters =
-    isNotNull(draftParameterValues) &&
-    !_.isEqual(parameterValues, draftParameterValues);
-
-  const handleApplyFilters = () => {
-    setParameterValues(draftParameterValues);
-  };
-
-  const applyFilterButton = !dashboard?.auto_apply_filters && (
-    <ApplyButton
-      primary
-      isVisible={hasUnappliedFilters}
-      onClick={handleApplyFilters}
-    >
-      Apply
-    </ApplyButton>
-  );
-
   return (
     <div className="shrink-below-content-size full-height">
       <Dashboard
         editingOnLoad={editingOnLoad}
         addCardOnLoad={addCardOnLoad}
-        applyFilterButton={applyFilterButton}
         {...props}
       />
       {/* For rendering modal urls */}
