+++ frontend/src/metabase/dashboard/hoc/DashboardControls.jsx
@@ -3,10 +3,13 @@ import React, { Component } from "react";
 
 import { connect } from "react-redux";
 import { replace } from "react-router-redux";
+import _ from "underscore";
 
 import screenfull from "screenfull";
 import * as MetabaseAnalytics from "metabase/lib/analytics";
 import { parseHashOptions, stringifyHashOptions } from "metabase/lib/browser";
+import { isNotNull } from "metabase/core/utils/types";
+import { ApplyButton } from "./DashboardControls.styled";
 
 const TICK_PERIOD = 1; // seconds
 
@@ -213,6 +216,29 @@ export default ComposedComponent =>
       };
 
       render() {
+        const {
+          dashboard,
+          parameterValues,
+          draftParameterValues,
+          setParameterValues,
+        } = this.props;
+        const hasUnappliedFilters =
+          isNotNull(draftParameterValues) &&
+          !_.isEqual(parameterValues, draftParameterValues);
+
+        const handleApplyFilters = () => {
+          setParameterValues(draftParameterValues);
+        };
+
+        const applyFilterButton = !dashboard?.auto_apply_filters ? (
+          <ApplyButton
+            primary
+            isVisible={hasUnappliedFilters}
+            onClick={handleApplyFilters}
+          >
+            Apply
+          </ApplyButton>
+        ) : null;
         return (
           <ComposedComponent
             {...this.props}
@@ -225,6 +251,7 @@ export default ComposedComponent =>
             onNightModeChange={this.setNightMode}
             onFullscreenChange={this.setFullscreen}
             onRefreshPeriodChange={this.setRefreshPeriod}
+            applyFilterButton={applyFilterButton}
           />
         );
       }
